requirements.txt

flask
werkzeug

sqlalchemy
psycopg2-binary
python-dotenv
